adding start index
